Gjort slik at karakteren kan kun gå til "midten" av tiles når man klikker. For øyeblikket sjekker ikke denne om spilleren går ut av mappen, men det går til det som ville vært midten hvis en tile var "rendret" der.

+++ src/main/java/inf112/saga/of/the/villeins/Controller/CharacterAnimationController.java
@@ -43,10 +43,10 @@ public class CharacterAnimationController {
 
 
         if (rows != null) {
-            idleframeRows = rows.intValue();
+            idleframeRows = rows;
         }
         if (rows != null) {
-            idleFrameCols = cols.intValue();
+            idleFrameCols = cols;
         }
